+++ web/src/actions/app/symbols.js
@@ -1,3 +1,15 @@
+export const CREATE_APD_FAILURE = Symbol('create apd failure');
+export const CREATE_APD_REQUEST = Symbol('create apd request');
+export const CREATE_APD_SUCCESS = Symbol('create apd success');
+
+export const DELETE_APD_FAILURE = Symbol('delete apd failure');
+export const DELETE_APD_REQUEST = Symbol('delete apd request');
+export const DELETE_APD_SUCCESS = Symbol('delete apd success');
+
+export const FETCH_ALL_APDS_FAILURE = Symbol('fetch all apds failure');
+export const FETCH_ALL_APDS_REQUEST = Symbol('fetch all apds request');
+export const FETCH_ALL_APDS_SUCCESS = Symbol('fetch all apds success');
+
 export const SAVE_APD_FAILURE = Symbol('save apd failure');
 export const SAVE_APD_REQUEST = Symbol('save apd request');
 export const SAVE_APD_SUCCESS = Symbol('save apd success');
